Fresh Graphics

+++ src/main/java/net/judah/effects/api/Effect.java
@@ -5,8 +5,12 @@ public interface Effect {
     String getName();
     boolean isActive();
     void setActive(boolean active);
-    void set(int idx, float value);
-    float get(int idx);
+    /**@param idx
+     * @param value scaled from 0 to 100 */
+    void set(int idx, int value);
+    /**@param idx
+     * @return parameter value of idx scaled from 0 to 100*/
+    int get(int idx);
     int getParamCount();
-
+    
 }
